+++ src/dipdup/context.py
@@ -51,6 +51,7 @@ from dipdup.exceptions import CallbackError
 from dipdup.exceptions import CallbackTypeError
 from dipdup.exceptions import ConfigurationError
 from dipdup.exceptions import ContractAlreadyExistsError
+from dipdup.exceptions import DipDupError
 from dipdup.exceptions import InitializationRequiredError
 from dipdup.exceptions import ReindexingRequiredError
 from dipdup.models import Contract
@@ -206,7 +207,11 @@ class DipDupContext:
         elif action == ReindexingAction.wipe:
             conn = get_connection()
             if isinstance(self.config.database, PostgresDatabaseConfig):
-                await wipe_schema(conn, self.config.database.schema_name, self.config.database.immune_tables)
+                await wipe_schema(
+                    conn=conn,
+                    schema_name=self.config.database.schema_name,
+                    immune_tables=self.config.database.immune_tables,
+                )
             else:
                 await Tortoise._drop_databases()
             await self.restart()
@@ -601,7 +606,8 @@ class CallbackManager:
                     stack.enter_context(Metrics.measure_callback_duration(module))
                 yield
         except Exception as e:
-            if isinstance(e, ReindexingRequiredError):
+            # NOTE: Do not wrap known errors like ProjectImportError
+            if isinstance(e, DipDupError):
                 raise
             raise CallbackError(module, e) from e