+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog],
 and this project adheres to [Semantic Versioning].
 
+## [6.0.1] - 2022-08-19
+
+### Fixed
+
+- codegen: Fixed invalid `models.py` template.
+- context: Do not wrap known exceptions with `CallbackError`.
+- database: Raise `DatabaseConfigurationError` when backward relation name equals table name.
+- database: Wrap schema wiping in a transaction to avoid orphaned tables in the immune schema.
+- hasura: Fixed processing M2M relations.
+- sentry: Fixed "invalid value `environment`" error.
+- sentry: Ignore events from project callbacks when `crash_reporting` is enabled.
+
 ## [6.0.0] - 2022-08-08
 
 This release contains no changes except for the version number.
@@ -693,7 +705,8 @@ This release contains no changes except for the version number.
 [semantic versioning]: https://semver.org/spec/v2.0.0.html
 
 <!-- Versions -->
-[unreleased]: https://github.com/dipdup-net/dipdup-py/compare/6.0.0...HEAD
+[Unreleased]: https://github.com/dipdup-net/dipdup-py/compare/6.0.1...HEAD
+[6.0.1]: https://github.com/dipdup-net/dipdup-py/compare/6.0.0...6.0.1
 [6.0.0]: https://github.com/dipdup-net/dipdup-py/compare/6.0.0rc2...6.0.0
 [6.0.0rc2]: https://github.com/dipdup-net/dipdup-py/compare/6.0.0-rc1...6.0.0rc2
 [6.0.0-rc1]: https://github.com/dipdup-net/dipdup-py/compare/5.2.5...6.0.0-rc1